FOUR-20583-A: Tasks Main change to new layout
