+++ resources/js/next/monaco.js
@@ -1,19 +1,19 @@
-import MonacoEditor from "vue-monaco";
+// import MonacoEditor from "vue-monaco";
 import { getGlobalVariable, setGlobalVariable } from "./globalVariables";
-
-const Vue = getGlobalVariable("Vue");
-
-Vue.component("MonacoEditor", MonacoEditor);
-setGlobalVariable("VueMonaco", MonacoEditor);
-setGlobalVariable("monaco", MonacoEditor);
 
 // const Vue = getGlobalVariable("Vue");
 
-// Vue.component("MonacoEditor", (resolve, reject) => {
-//   console.log("LOADER MONACO %%%%%%%%%%");
+// Vue.component("MonacoEditor", MonacoEditor);
+// setGlobalVariable("VueMonaco", MonacoEditor);
+// setGlobalVariable("monaco", MonacoEditor);
 
-//   import("vue-monaco").then((MonacoEditor) => {
-//     setGlobalVariable("VueMonaco", MonacoEditor.default);
-//     resolve(MonacoEditor.default);
-//   }).catch(reject);
-// });
+const Vue = getGlobalVariable("Vue");
+
+Vue.component("MonacoEditor", (resolve, reject) => {
+  console.log("LOADER MONACO %%%%%%%%%%");
+
+  import("vue-monaco").then((MonacoEditor) => {
+    setGlobalVariable("VueMonaco", MonacoEditor.default);
+    resolve(MonacoEditor.default);
+  }).catch(reject);
+});
